fixed all gcc warnings except 2 concerning fprintf

+++ src/bitscan/bbscan_sparse.h
@@ -251,7 +251,7 @@ namespace bitgraph {
 		else {
 
 			//Searches for next bit in the remaining blocks
-			for (auto i = scan_.bbi_ + 1; i < vBB_.size(); ++i) {
+			for (auto i = scan_.bbi_ + 1; i < (int)	vBB_.size(); ++i) {
 				if (_BitScanForward64(&posInBB, vBB_[i].bb_)) {
 					//stores the current block and bit for next call
 					scan_.bbi_ = i;
